Analyze all databases with vacuumdb
When using the "logical" databases initialization,
we need to analyze our target database (that one that
has the data we just loaded).

It looks like conn.DBName is pointing to "postgres"
when we run vacuumdb.

Switch to using "-all" for running ANALYZE. The most time is
to be spent in our target database since only that database has
significant volumes of data.